Removed NullAI, AIFactory and AIWrapper. Renamed Keeper2AI to KeeperAI Most of the AIWrapper has been moved to BaseAI (to be shared between AIs) The factory has been moved to the AIManager fixes #328
