--- source/ai/AIWrapper.cpp
@@ -1,556 +1,0 @@
-/*!
- *  Copyright (C) 2011-2014  OpenDungeons Team
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "ai/AIWrapper.h"
-
-#include "game/Seat.h"
-#include "game/Player.h"
-#include "gamemap/GameMap.h"
-#include "entities/Creature.h"
-#include "network/ODServer.h"
-#include "utils/LogManager.h"
-
-const int32_t pointsPerWallSpot = 50;
-const int32_t handicapPerTileOffset = 20;
-
-AIWrapper::AIWrapper(GameMap& gameMap, Player& player)
-    : gameMap(gameMap), player(player), seat(*player.getSeat())
-{
-}
-
-AIWrapper::~AIWrapper()
-{
-}
-
-bool AIWrapper::buildRoom(Room* room, const std::vector<Tile*>& tiles)
-{
-    if(!ODServer::getSingleton().isConnected())
-        return false;
-
-    room->setupRoom(gameMap.nextUniqueNameRoom(room->getMeshName()), player.getSeat(), tiles);
-    gameMap.addRoom(room);
-    room->checkForRoomAbsorbtion();
-    room->createMesh();
-    room->updateActiveSpots();
-    gameMap.refreshBorderingTilesOf(tiles);
-
-    return true;
-}
-
-bool AIWrapper::dropHand(int x, int y, int index)
-{
-    if(!player.isDropHandPossible(gameMap.getTile(x, y), index))
-        return false;
-
-    player.dropHand(gameMap.getTile(x, y), index);
-    return true;
-}
-
-bool AIWrapper::pickUpCreature(Creature* creature)
-{
-    player.pickUpEntity(creature, false);
-    return true;
-}
-
-const std::vector<GameEntity*>& AIWrapper::getObjectsInHand()
-{
-    return player.getObjectsInHand();
-}
-
-std::vector<const Room*> AIWrapper::getOwnedRoomsByType(Room::RoomType type)
-{
-    const GameMap& gm = gameMap;
-    return gm.getRoomsByTypeAndSeat(type, &seat);
-}
-
-Room* AIWrapper::getDungeonTemple()
-{
-    std::vector<Room*> dt = gameMap.getRoomsByTypeAndSeat(Room::dungeonTemple, &seat);
-    if(!dt.empty())
-        return dt.front();
-    else
-        return NULL;
-}
-
-void AIWrapper::markTileForDigging(Tile* tile)
-{
-    tile->setMarkedForDigging(true, &player);
-}
-
-int AIWrapper::getGoldInTreasury() const
-{
-    return seat.getGold();
-}
-
-std::vector<Tile*> AIWrapper::getAffectedTiles(int x1, int y1, int x2, int y2)
-{
-    std::vector<Tile*> affectedTiles = gameMap.rectangularRegion(x1, y1, x2, y2);
-    std::vector<Tile*>::iterator it = affectedTiles.begin();
-    while(it != affectedTiles.end())
-    {
-        if((*it)->getSeat() != &seat || !(*it)->isBuildableUpon())
-        {
-            it = affectedTiles.erase(it);
-        }
-        else
-        {
-            ++it;
-        }
-    }
-    return affectedTiles;
-}
-
-bool AIWrapper::shouldGroundTileBeConsideredForBestPlaceForRoom(Tile* tile, Seat* playerSeat)
-{
-    switch(tile->getType())
-    {
-        case Tile::TileType::dirt:
-        case Tile::TileType::gold:
-        {
-            // Dirt and gold can always be built (even if digging may be needed depending on fullness)
-            return true;
-        }
-        case Tile::TileType::claimed:
-        {
-            // We check if we can build on that tile and if there is no building currently
-            if(!tile->isClaimedForSeat(playerSeat))
-                return false;
-            if(tile->getCoveringRoom() != nullptr)
-                return false;
-            if(tile->getCoveringTrap() != nullptr)
-                return false;
-
-            // We don't want to break a wall where there are activespots from another one
-            for(Tile* t : tile->getAllNeighbors())
-            {
-                if(t->isClaimedForSeat(playerSeat) &&
-                    t->getCoveringRoom() != nullptr)
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-        default:
-            return false;
-    }
-
-    return false;
-}
-
-bool AIWrapper::shouldWallTileBeConsideredForBestPlaceForRoom(Tile* tile, Seat* playerSeat)
-{
-    // We only consider wall claimed for the correct seat or dirt (that can be claimed)
-    if(tile->getFullness() <= 0.0)
-        return false;
-
-    if(tile->getType() == Tile::TileType::dirt)
-        return true;
-
-    if(tile->isWallClaimedForSeat(playerSeat))
-        return true;
-
-    return false;
-}
-
-//! To find the position, we try every square of the wantedSize width around the given tile for each possible distance
-bool AIWrapper::findBestPlaceForRoom(Tile* tile, Seat* playerSeat, int32_t wantedSize, bool useWalls,
-    int32_t& bestX, int32_t& bestY)
-{
-    // We use a point system to find the best position. Once we find a valid position, we will set a handicap
-    // that will increase as we go away from the given tile. Once the handicap is > to the max points we can get minus
-    // the points the room we found got, we can stop searching.
-    // With this logic, we can tune easily what the AI should prefer between distance and active spots.
-
-    // We search for the maximum points a room can get
-    int32_t maxPointsPossible = 0;
-    if(wantedSize >= 3)
-    {
-        // Maximum central active spots
-        int32_t nbCentralActiveSpots = ((wantedSize - 3) / 2) + 1;
-        // Wall active spots
-        if(useWalls)
-            maxPointsPossible += nbCentralActiveSpots * 4 * pointsPerWallSpot;
-    }
-
-    bool isFound = false;
-    int32_t handicap = 0;
-    int32_t bestPoints = 0;
-    int32_t bestDistance = 0;
-    int32_t maxOffset = std::max(gameMap.getMapSizeX(), gameMap.getMapSizeY());
-
-    for(int32_t offset = 1; offset < maxOffset; ++offset)
-    {
-        int32_t points = 0;
-        int32_t nbTiles = offset * 2 + wantedSize - 1;
-        for(int32_t k = 0; k < nbTiles; ++k)
-        {
-            Tile* t;
-            // North
-            t  = gameMap.getTile(tile->getX() - offset - wantedSize + 2 + k, tile->getY() + offset);
-            if((t != nullptr) &&
-               computePointsForRoom(t, playerSeat, wantedSize, true, useWalls, points))
-            {
-                points -= handicap;
-                int32_t centerX = t->getX() + (wantedSize / 2);
-                int32_t centerY = t->getY() + (wantedSize / 2);
-                int32_t distance = (tile->getX() - centerX) * (tile->getX() - centerX);
-                distance += (tile->getY() - centerY) * (tile->getY() - centerY);
-                if((points > bestPoints) ||
-                   (points == bestPoints && distance < bestDistance))
-                {
-                    bestDistance = distance;
-                    bestX = t->getX();
-                    bestY = t->getY();
-                    bestPoints = points;
-                    isFound = true;
-                }
-            }
-            // East
-            t  = gameMap.getTile(tile->getX() + offset, tile->getY() - k + offset);
-            if((t != nullptr) &&
-               computePointsForRoom(t, playerSeat, wantedSize, true, useWalls, points))
-            {
-                points -= handicap;
-                int32_t centerX = t->getX() + (wantedSize / 2);
-                int32_t centerY = t->getY() + (wantedSize / 2);
-                int32_t distance = (tile->getX() - centerX) * (tile->getX() - centerX);
-                distance += (tile->getY() - centerY) * (tile->getY() - centerY);
-                if((points > bestPoints) ||
-                   (points == bestPoints && distance < bestDistance))
-                {
-                    bestDistance = distance;
-                    bestX = t->getX();
-                    bestY = t->getY();
-                    bestPoints = points;
-                    isFound = true;
-                }
-            }
-            // South
-            t  = gameMap.getTile(tile->getX() + offset + wantedSize - 2 - k, tile->getY() - offset);
-            if((t != nullptr) &&
-               computePointsForRoom(t, playerSeat, wantedSize, false, useWalls, points))
-            {
-                points -= handicap;
-                int32_t centerX = t->getX() - (wantedSize / 2);
-                int32_t centerY = t->getY() - (wantedSize / 2);
-                int32_t distance = (tile->getX() - centerX) * (tile->getX() - centerX);
-                distance += (tile->getY() - centerY) * (tile->getY() - centerY);
-                if((points > bestPoints) ||
-                   (points == bestPoints && distance < bestDistance))
-                {
-                    bestDistance = distance;
-                    bestX = t->getX() - wantedSize + 1;
-                    bestY = t->getY() - wantedSize + 1;
-                    bestPoints = points;
-                    isFound = true;
-                }
-            }
-            // West
-            t  = gameMap.getTile(tile->getX() - offset, tile->getY() - offset + k);
-            if((t != nullptr) &&
-               computePointsForRoom(t, playerSeat, wantedSize, false, useWalls, points))
-            {
-                points -= handicap;
-                int32_t centerX = t->getX() - (wantedSize / 2);
-                int32_t centerY = t->getY() - (wantedSize / 2);
-                int32_t distance = (tile->getX() - centerX) * (tile->getX() - centerX);
-                distance += (tile->getY() - centerY) * (tile->getY() - centerY);
-                if((points > bestPoints) ||
-                   (points == bestPoints && distance < bestDistance))
-                {
-                    bestDistance = distance;
-                    bestX = t->getX() - wantedSize + 1;
-                    bestY = t->getY() - wantedSize + 1;
-                    bestPoints = points;
-                    isFound = true;
-                }
-            }
-        }
-
-        if(isFound)
-        {
-            handicap += handicapPerTileOffset;
-            // If we already found the best place, stop searching
-            if(handicap > (maxPointsPossible - bestPoints))
-                break;
-        }
-    }
-    return isFound;
-}
-
-bool AIWrapper::computePointsForRoom(Tile* tile, Seat* playerSeat, int32_t wantedSize,
-    bool bottomLeft2TopRight, bool useWalls, int32_t& points)
-{
-    int tileX = tile->getX();
-    int tileY = tile->getY();
-    bool isOk = true;
-
-    // We check if the tile is reachable. No need to compute if the tile is behind rocks
-    points = 0;
-    // We start by searching for a place fitting the min size. If we find it,
-    // we increase tile by tile until the max size. If not, no need to continue
-    for(int32_t xx = 0; xx < wantedSize; ++xx)
-    {
-        for(int32_t yy = 0; yy < wantedSize; ++yy)
-        {
-            Tile* t;
-            if(bottomLeft2TopRight)
-                t  = gameMap.getTile(tileX + xx, tileY + yy);
-            else
-                t  = gameMap.getTile(tileX - xx, tileY - yy);
-
-            if(t == nullptr)
-            {
-                isOk = false;
-                break;
-            }
-
-            if(!shouldGroundTileBeConsideredForBestPlaceForRoom(t, playerSeat))
-            {
-                isOk = false;
-                break;
-            }
-        }
-
-        if(!isOk)
-            break;
-    }
-
-    if(!isOk)
-        return false;
-
-    // If we don't want to consider walls, we stop here (for example for rooms that do not have bonus
-    // with wall active spots like treasury or dormitory)
-    if(!useWalls)
-        return true;
-
-    // We search points for first wall. That's not exactly how the activespots will be computed but it will be enough (especially
-    // when the room size is even)
-    int nbConsecutiveTiles;
-    int nbActiveWallSpots;
-    // We search points for first wall
-    nbConsecutiveTiles = 0;
-    nbActiveWallSpots = 0;
-    for(int32_t kk = 0; kk < wantedSize; ++kk)
-    {
-        Tile* t;
-        if(bottomLeft2TopRight)
-            t  = gameMap.getTile(tileX - 1, tileY + kk);
-        else
-            t  = gameMap.getTile(tileX + 1, tileY - kk);
-        if(t == nullptr)
-            continue;
-
-        if(shouldWallTileBeConsideredForBestPlaceForRoom(t, playerSeat))
-            ++nbConsecutiveTiles;
-        else
-            nbConsecutiveTiles = 0;
-
-        if(nbActiveWallSpots == 0)
-        {
-            if(nbConsecutiveTiles >= 3)
-            {
-                nbConsecutiveTiles = 0;
-                ++nbActiveWallSpots;
-            }
-        }
-        else if(nbConsecutiveTiles >= 2)
-        {
-            nbConsecutiveTiles = 0;
-            ++nbActiveWallSpots;
-        }
-    }
-    points += nbActiveWallSpots * pointsPerWallSpot;
-    // We search points for second wall
-    nbConsecutiveTiles = 0;
-    nbActiveWallSpots = 0;
-    for(int32_t kk = 0; kk < wantedSize; ++kk)
-    {
-        Tile* t;
-        if(bottomLeft2TopRight)
-            t = gameMap.getTile(tileX + wantedSize, tileY + kk);
-        else
-            t = gameMap.getTile(tileX - wantedSize, tileY - kk);
-        if(t == nullptr)
-            continue;
-
-        if(shouldWallTileBeConsideredForBestPlaceForRoom(t, playerSeat))
-            ++nbConsecutiveTiles;
-        else
-            nbConsecutiveTiles = 0;
-
-        if(nbActiveWallSpots == 0)
-        {
-            if(nbConsecutiveTiles >= 3)
-            {
-                nbConsecutiveTiles = 0;
-                ++nbActiveWallSpots;
-            }
-        }
-        else if(nbConsecutiveTiles >= 2)
-        {
-            nbConsecutiveTiles = 0;
-            ++nbActiveWallSpots;
-        }
-    }
-    points += nbActiveWallSpots * pointsPerWallSpot;
-    // We search points for first wall
-    nbConsecutiveTiles = 0;
-    nbActiveWallSpots = 0;
-    for(int32_t kk = 0; kk < wantedSize; ++kk)
-    {
-        Tile* t;
-        if(bottomLeft2TopRight)
-            t = gameMap.getTile(tileX + kk, tileY - 1);
-        else
-            t = gameMap.getTile(tileX - kk, tileY + 1);
-        if(t == nullptr)
-            continue;
-
-        if(shouldWallTileBeConsideredForBestPlaceForRoom(t, playerSeat))
-            ++nbConsecutiveTiles;
-        else
-            nbConsecutiveTiles = 0;
-
-        if(nbActiveWallSpots == 0)
-        {
-            if(nbConsecutiveTiles >= 3)
-            {
-                nbConsecutiveTiles = 0;
-                ++nbActiveWallSpots;
-            }
-        }
-        else if(nbConsecutiveTiles >= 2)
-        {
-            nbConsecutiveTiles = 0;
-            ++nbActiveWallSpots;
-        }
-    }
-    points += nbActiveWallSpots * pointsPerWallSpot;
-    // We search points for first wall
-    nbConsecutiveTiles = 0;
-    nbActiveWallSpots = 0;
-    for(int32_t kk = 0; kk < wantedSize; ++kk)
-    {
-        Tile* t;
-        if(bottomLeft2TopRight)
-            t = gameMap.getTile(tileX + kk, tileY + wantedSize);
-        else
-            t = gameMap.getTile(tileX - kk, tileY - wantedSize);
-        if(t == nullptr)
-            continue;
-
-        if(shouldWallTileBeConsideredForBestPlaceForRoom(t, playerSeat))
-            ++nbConsecutiveTiles;
-        else
-            nbConsecutiveTiles = 0;
-
-        if(nbActiveWallSpots == 0)
-        {
-            if(nbConsecutiveTiles >= 3)
-            {
-                nbConsecutiveTiles = 0;
-                ++nbActiveWallSpots;
-            }
-        }
-        else if(nbConsecutiveTiles >= 2)
-        {
-            nbConsecutiveTiles = 0;
-            ++nbActiveWallSpots;
-        }
-    }
-    points += nbActiveWallSpots * pointsPerWallSpot;
-
-    return true;
-}
-
-bool AIWrapper::digWayToTile(Tile* tileStart, Tile* tileEnd)
-{
-    // We find a way to tileEnd. We search in reverse order to stop when we reach the first
-    // accessible tile
-
-    // Set a diggable path up to the first gold spot for the given team color, by the first available kobold
-    Seat* seat = player.getSeat();
-    std::vector<Creature*> creatures = gameMap.getCreaturesBySeat(seat);
-
-    if (creatures.empty())
-        return false;
-
-    Creature* kobold = nullptr;
-    for (Creature* creature : creatures)
-    {
-        if (creature->getDefinition()->getClassName() == "Kobold")
-        {
-            kobold = creature;
-            break;
-        }
-    }
-
-    if (kobold == nullptr)
-        return false;
-
-    std::list<Tile*> pathToDig = gameMap.path(tileEnd, tileStart, kobold,
-        seat, true);
-    if (pathToDig.empty())
-        return false;
-
-    // We search for the first reachable tile in the list
-    bool isPathFound = false;
-    for(std::list<Tile*>::iterator it = pathToDig.begin(); it != pathToDig.end();)
-    {
-        Tile* tile = *it;
-        if(!isPathFound &&
-           (tile->getFullness() == 0.0) &&
-           (gameMap.pathExists(kobold, tileStart, tile)))
-        {
-            isPathFound = true;
-        }
-
-        if(isPathFound)
-        {
-            it = pathToDig.erase(it);
-            continue;
-        }
-
-        // If the tile should be dug, we check if one of its neighboors can be reached.
-        // If yes, we will stop after digging it to avoid digging through a wall as much as
-        // possible
-        for(Tile* t : tile->getAllNeighbors())
-        {
-            if(!isPathFound &&
-               (t->getFullness() == 0.0) &&
-               (gameMap.pathExists(kobold, tileStart, t)))
-            {
-                // we let the iterator increment because we want to dig the currently
-                // tested tile
-                isPathFound = true;
-                break;
-            }
-        }
-
-        ++it;
-    }
-
-    for(Tile* tile : pathToDig)
-    {
-        if (tile && tile->isDiggable(seat))
-            markTileForDigging(tile);
-    }
-
-    return true;
-}
